[AL-10894] Move new stage to destination
