+++ agent/src/agent/pipeline/config/stages/destination.py
@@ -26,6 +26,33 @@ class WatermarkDestination(Stage):
         }
 
 
+class WatermarkWithMetricsDestination(JythonProcessor):
+    JYTHON_SCRIPT = 'send_watermark_and_monitoring_metrics.py'
+
+    def _get_script_params(self) -> list[dict]:
+        return [
+            {
+                'key': 'PROXIES',
+                'value': proxy.get_config(self.pipeline.destination.proxy)
+            },
+            {
+                'key': 'WATERMARK_URL',
+                'value': urllib.parse.urljoin(
+                    self.pipeline.streamsets.agent_external_url, f'/pipelines/{self.pipeline.name}/watermark')
+            },
+            {
+                'key': 'WATERMARK_DELTA_MONITORING_ENDPOINT',
+                'value': urllib.parse.urljoin(
+                    self.pipeline.streamsets.agent_external_url, f'/monitoring/watermark_delta/{self.pipeline.name}')
+            },
+            {
+                'key': 'WATERMARK_SENT_MONITORING_ENDPOINT',
+                'value': urllib.parse.urljoin(
+                    self.pipeline.streamsets.agent_external_url, f'/monitoring/watermark_sent/{self.pipeline.name}')
+            },
+        ]
+
+
 class EventsDestination(Stage):
     def get_config(self) -> dict:
         return self.pipeline.destination.config
